nvme-types: move nvm definitions into separate header
The nvme-types.h header has become too large and difficult to maintain.
