+++ libnvme/src/nvme/nvme-cmds-nvm.h
@@ -20,7 +20,7 @@
 
 #include <nvme/endian.h>
 #include <nvme/ioctl.h>
-#include <nvme/nvme-types.h>
+#include <nvme/nvme-types-nvm.h>
 #include <nvme/nvme-cmds-base.h>
 
 /**
